docs(pull-down-refresh): update docs and demo
