+++ src/pull-down-refresh/_example/base.jsx
@@ -1,50 +1,15 @@
 import React from 'react';
-import { PullDownRefresh, Skeleton } from 'tdesign-mobile-react';
-import './style/index.less';
+import { PullDownRefresh } from 'tdesign-mobile-react';
 
-export default function Demo() {
-  const rowCols = [
-    {
-      height: '171px',
-      borderRadius: '8px',
-    },
-    [
-      {
-        width: '50%',
-        type: 'circle',
-      },
-      {
-        width: '50%',
-        type: 'circle',
-      },
-    ],
-    [
-      {
-        width: '30%',
-        type: 'circle',
-      },
-      {
-        width: '30%',
-        type: 'circle',
-      },
-    ],
-  ];
+export default function BaseDemo({ children }) {
   return (
     <div className="tdesign-mobile-wrapper">
-      <div style={{ display: 'flex' }}>
-        <div className="tdesign-mobile-item">
-          <Skeleton theme="text" rowCol={rowCols} />
-        </div>
-        <div className="tdesign-mobile-item">
-          <Skeleton theme="text" rowCol={rowCols} />
-        </div>
-      </div>
-
       <PullDownRefresh
         loadingBarHeight={80}
         loadingProps={{
           layout: 'vertical',
         }}
+        loadingTexts={['下拉刷新', '松开刷新', '正在刷新', '刷新完成']}
         onRefresh={() =>
           new Promise((resolve) => {
             setTimeout(() => {
@@ -53,7 +18,7 @@ export default function Demo() {
           })
         }
       >
-        <div className="tdesign-mobile-block">拖拽该区域演示 中间下拉刷新</div>
+        {children}
       </PullDownRefresh>
     </div>
   );
