+++ src/pull-down-refresh/_example/style/index.less
@@ -7,25 +7,21 @@
 .tdesign-mobile-demo {
   background: #fff;
 
-  .tdesign-mobile-wrapper {
-    padding: 0 12px;
-  }
-
   .tdesign-mobile-block {
     box-sizing: border-box;
-    background-color: #eee;
+    background-color: rgba(243, 243, 243, 1);;
     border-radius: 8px;
-    width: 100%;
     height: 171px;
     font-size: 16px;
     color: rgba(0, 0, 0, 0.26);
     line-height: 171px;
     text-align: center;
-    margin-bottom: 12px;
+    margin: 16px;
   }
 
   .tdesign-mobile-item {
     width: 48%;
+    margin: 16px;
 
     &:last-child {
       margin-left: 4%;
